fix broken ITs and refactored the tests
